backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
prometheus-fastapi-instrumentator
jinja2